requirements.txt
feedparser
feedgen
python-dateutil
pytz